backend/requirements.txt
alembic==1.13.1
bidict==0.23.1
blinker==1.8.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.3
pandas==2.2.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.2
pytz==2024.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
tzdata==2024.1
Werkzeug==3.0.3
wsproto==1.2.0